+++ src/query.rs
@@ -808,7 +808,7 @@ where
     #[cfg(feature = "par-iter")]
     pub fn par_entities_for_each<'a, T>(&'a mut self, world: &'a World, f: T)
     where
-        T: Fn((Entity, <<V as View<'_>>::Iter as std::iter::Iterator>::Item)) + Send + Sync,
+        T: Fn((Entity, <<V as View<'a>>::Iter as Iterator>::Item)) + Send + Sync,
     {
         self.par_iter_chunks(world).for_each(|mut chunk| {
             for data in chunk.iter_entities() {
